call repositionItems() in topic.edit and also in topic.show
